third_party/blink/web_tests/webaudio/OfflineAudioContext/offlineaudiocontext-thread-smoke-test.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      OfflineAudioContext - Thread Smoke Test
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      // The common sample rate.
      const sampleRate = 48000;

      // To make channel count and buffer length irrelevant to this test.
      const numberOfChannels = 1;
      const renderLength = 1;

      // 1000 concurrent contexts caused this test to time out on Mac OS and
      // Linux. (See: crbug.com/1092145, crbug.com/741699)
      const maxNumberOfContexts = 500;

      // Create and sequentially render multiple OfflineAudioContexts to verify
      // that no out-of-thread errors or crashes occur.
      promise_test(async () => {
        const contexts = [];

        // Phase 1: Create all contexts.
        for (let i = 0; i < maxNumberOfContexts; ++i) {
          contexts.push(new OfflineAudioContext(
              numberOfChannels, renderLength, sampleRate));
        }
        assert_equals(
            contexts.length, maxNumberOfContexts,
            'The correct number of contexts should be created without a crash');

        // Phase 2: Render all contexts sequentially.
        let renderedContexts = 0;
        for (const context of contexts) {
          await context.startRendering();
          renderedContexts++;
        }
        assert_equals(
            renderedContexts, maxNumberOfContexts,
            'All created contexts should render sequentially without a crash');
      }, `Create and sequentially render ${maxNumberOfContexts} ` +
          `OfflineAudioContexts.`);
    </script>
  </body>
</html>
